pyproject.toml
[project]
name = "outrl"
version = "0.1.0"
description = "An outside the box framework for training agents"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
  {name = "K.R. Zentner", email = "krzentner@gmail.com" }
]

classifiers = [
  "Development Status :: 3 - Alpha",

  "Intended Audience :: Science/Research",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3 :: Only",
]

# psutil should probably be at least 5.1.0
# If the user has an older version than that there's probably a good reason.
dependencies = [
  "psutil"
]

[project.optional-dependencies]
dev = []
test = ["pytest"]

[project.urls]
"Homepage" = "https://github.com/krzentner/outrl"
"Bug Reports" = "https://github.com/krzentner/outrl/issues"
"Source" = "https://github.com/krzentner/outrl"

# Below is poetry setup for developing this package
[tool.poetry.dependencies]
python = "^3.8,<3.11"
psutil = "^5.1.0"
torch = "^2.0"
lightning = "^2.0.3"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
gym = "0.17.3"
ipdb = "^0.13.13"

[tool.pyright]
include = ["src"]
venv = ".venv"

[tool.poetry]
name = "outrl"
version = "0.1.0"
description = ""
authors = ["K.R. Zentner <krzentner@gmail.com>"]
